refactor:packSimulation 분리
