+++ frontend/src/App.tsx
@@ -1,50 +1,16 @@
-import { useCallback, useState } from 'react';
-
+import Article from './components/Article/Article';
 import { Global } from '@emotion/react';
 import MainLayout from './layouts/MainLayout/MainLayout';
-import PackOpen from './components/PackOpen/PackOpen';
-import PackSelect from './components/PackSelect/PackSelect';
-import getRandomPack from './utils/getRandomPack';
+import PackSimulator from './components/PackSimulator/PackSimulator';
 import reset from './reset.style';
 
-const initPack = getRandomPack();
-
-type Phase = 'select' | 'open' | 'result';
 function App() {
-  const [cardPack, setCardPack] = useState<CardInfo[]>(initPack);
-  const [nowPackType, setNowPackType] = useState<A1PackType>('charizard');
-  const [phase, setPhase] = useState<Phase>('select');
-  const goOpenPhase = useCallback((packType: A1PackType) => {
-    const randomPack = getRandomPack(packType);
-    setCardPack(randomPack);
-    setPhase('open');
-    setNowPackType(packType);
-  }, []);
-
-  const reopen = () => {
-    const randomPack = getRandomPack(nowPackType);
-    setCardPack(randomPack);
-  };
-
-  const goSelect = () => {
-    setPhase('select');
-  };
   return (
     <>
       <Global styles={reset} />
       <MainLayout>
-        {/* {cardPack.map(c => c.cardName).join(' ')} */}
-        {phase === 'select' && (
-          <PackSelect onSelect={goOpenPhase} startPackType={nowPackType} />
-        )}
-        {phase === 'open' && (
-          <PackOpen
-            cardInfos={cardPack}
-            goOpen={reopen}
-            goSelect={goSelect}
-            nowPackType={nowPackType}
-          />
-        )}
+        <PackSimulator />
+        <Article />
       </MainLayout>
     </>
   );
